moved delete to MoodController
