+++ MoodSpace/app/src/main/java/com/example/moodspace/MoodCallbackId.java
@@ -1,0 +1,6 @@
+package com.example.moodspace;
+
+public enum MoodCallbackId implements CallbackId {
+    MOOD_DELETE_SUCCESS,
+    MOOD_DELETE_FAIL;
+}
